Modified a single source file, decetable6.cpp
Corrected the increment of variable idx; it was being incremented
before the innermost loop, corresponding to the secondary particle
energy, was being printed. This means that the next idx energy/cross
section values were being printed next to the current angle.

Moving the idx++ after calls to outVal in the i2=0 to nep-1 loop fixes
it.